\docType{methods}
\name{calcIRDV01-method}
\alias{calcIRDV01,CDS-method}
\alias{calcIRDV01-method}
\title{The calcIRDV01 method for CDS class. Calculate the IR DV01 from conventional spread}
\usage{
\S4method{calcIRDV01}{CDS}(object)
}
\arguments{
  \item{object}{the input CDS class object}
}
\description{
The calcIRDV01 method for CDS class. Calculate the IR DV01
from conventional spread
}

