\docType{methods}
\name{calcRecRisk01-method}
\alias{calcRecRisk01,CDS-method}
\alias{calcRecRisk01-method}
\title{Calculate the amount of change in upfront when there is a 1%
increase in recovery rate.}
\usage{
\S4method{calcRecRisk01}{CDS}(object)
}
\arguments{
  \item{object}{the input CDS class object}
}
\description{
Calculate the amount of change in upfront when there is a
1% increase in recovery rate.
}

