\docType{methods}
\name{calcSpreadDV01-method}
\alias{calcSpreadDV01,CDS-method}
\alias{calcSpreadDV01-method}
\title{The calcIRDV01 method for CDS class. Calculate the spread DV01 from conventional spread}
\usage{
\S4method{calcSpreadDV01}{CDS}(object)
}
\arguments{
  \item{object}{the input CDS class object}
}
\description{
The calcIRDV01 method for CDS class. Calculate the spread
DV01 from conventional spread
}

